requirements.txt
requests>=2.30.0
pandas>=2.0.0
networkx>=3.0
pyvis>=0.3.0
typer>=0.9.0
rich>=13.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
